hcore/serializers.py
from rest_framework import serializers
from .models import Account, Member


class MemberSerializer(serializers.ModelSerializer):
    class Meta:
        model = Member
        fields = '__all__'
        read_only_fields = ("id", "account_id")


class AccountSerializer(serializers.ModelSerializer):
    class Meta:
        model = Account
        fields = '__all__'
    read_only_fields = ("id",)
